refactor: rename `AuthClient` to `AuthApi`
